requirements.txt
# Z-Vision Dependencies
# 플랫폼에 따라 필요한 패키지만 설치하세요

# ============================================================
# 공통 (필수)
# ============================================================
pillow>=10.0.0
gradio>=4.0.0
huggingface-hub>=0.20.0

# ============================================================
# Apple Silicon (MLX) - macOS M1/M2/M3/M4
# pip install -r requirements.txt && pip install mflux
# ============================================================
# mflux>=0.13.0  # Apple Silicon 전용, 별도 설치 필요

# ============================================================
# CUDA / MPS / CPU (PyTorch)
# pip install -r requirements.txt
# ============================================================
torch>=2.0.0
diffusers>=0.31.0
transformers>=4.40.0
accelerate>=0.25.0
